cleaning up etsy-specific stuff
