--- forever-config.example.json
@@ -1,7 +1,0 @@
-{
-    "uid": "emojitools",
-    "script": "server.js",
-    "sourceDir": "/path/to/emojitools",
-    "watch": "true",
-    "append": "true"
-}
