+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "emoji-tools",
+  "name": "emojitools",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
